+++ formula1/src/main/java/xxx/joker/apps/formula1/webParser/WikiParser.java
@@ -8,8 +8,8 @@ public interface WikiParser {
 
     void parse();
 
-    Map<String, Integer> getExpectedDriverPoints();
-    Map<String, Integer> getExpectedTeamPoints();
+    Map<String, Double> getExpectedDriverPoints();
+    Map<String, Double> getExpectedTeamPoints();
 
     static WikiParser getParser(int year) {
         return JkReflection.createInstanceSafe("xxx.joker.apps.formula1.webParser.years.Year"+year);
